Quinto commit. Mejoras menores. Login y logout solo visibles segun autenticacion

+++ src/main/java/com/nombreGrupo/repositories/VerificacionUuidRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface VerificacionUuidRepository extends JpaRepository<VerificacionUuid, Long> {
+public interface VerificacionUuidRepository extends JpaRepository<VerificacionUuid, Integer> {
     Optional<VerificacionUuid> findByUuid(String UuidString);
     Optional<VerificacionUuid> findByUsuario_IdUsuario(int idUsuario);
 }
